+++ src/server/devices/shelly/index.ts
@@ -1,2 +1,3 @@
 export { ShellyDevice } from './shellyDevice';
 export { ShellyTrv } from './shellyTrv';
+export { ShellyActuator } from './shellyActuator';