+++ CHANGELOG.md
@@ -5,8 +5,7 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
-
-## **WORK IN PROGRESS**
+## 3.0.0-alpha.56 (2024-05-09)
 
 * (theimo1221) Add settings to prevent/allow Dachs to run/start based on battery level.
 * (theimo1221) Add logic and settings to control the Dachs WW pump.